feat: generate descriptions with metadata
Appends a paths metadata to each of its methods description.
Relies on an object in apikana-defaults, "metadata", containing the
properties "matchProps" and "mappedKeys".
"matchProps" expects a string containing a regular expression matching
the paths properties which should be handled as metadata, the default
being "^x-.*$".
"mappedKeys" expects an object containing mappings for the keys contained
in a metadata property, intended to make the descriptions better looking,
the default is an empty object.